+++ domains/list.go
@@ -5,9 +5,18 @@ import (
 	"os"
 
 	"github.com/Scalingo/cli/config"
+	scalingo "github.com/Scalingo/go-scalingo"
 	"github.com/olekukonko/tablewriter"
 	"gopkg.in/errgo.v1"
 )
+
+var letsencryptStatusString = map[string]string{
+	string(scalingo.LetsEncryptStatusPendingDNS):  "Pending DNS",
+	string(scalingo.LetsEncryptStatusNew):         "Creating",
+	string(scalingo.LetsEncryptStatusCreated):     "In use",
+	string(scalingo.LetsEncryptStatusDNSRequired): "DNS required",
+	string(scalingo.LetsEncryptStatusError):       "Error",
+}
 
 func List(app string) error {
 	c := config.ScalingoClient()
@@ -17,7 +26,7 @@ func List(app string) error {
 	}
 
 	t := tablewriter.NewWriter(os.Stdout)
-	t.SetHeader([]string{"Domain", "SSL"})
+	t.SetHeader([]string{"Domain", "TLS/SSL"})
 	hasCanonical := false
 
 	for _, domain := range domains {
@@ -29,6 +38,12 @@ func List(app string) error {
 		row := []string{domainName}
 		if !domain.SSL {
 			row = append(row, "-")
+		} else if domain.LetsEncrypt {
+			letsencryptStatus, ok := letsencryptStatusString[string(domain.LetsEncryptStatus)]
+			if !ok {
+				letsencryptStatus = string(domain.LetsEncryptStatus)
+			}
+			row = append(row, fmt.Sprintf("Let's Encrypt: %s", letsencryptStatus))
 		} else {
 			row = append(row, fmt.Sprintf("Valid until %v", domain.Validity))
 		}