+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### To be Released
 
 * [alerts] Add support for the duration_before_trigger attribute [#407](https://github.com/Scalingo/cli/pull/407)
+* [domains] Handle Let's Encrypt certificate status [#410](https://github.com/Scalingo/cli/pull/410)
 
 ### 1.10.1